doc: API documentation refactoring.
Restructured the API documentation and moved it from the Reference
section to the Developer guides section.

Added a note about the search function to help users find the API they
are looking for faster.

Splitted the API docs into two files one for nanokernel and another for
the microkernel APIs.

Added table of contents for each of the API sections with a back link on
the headings to take the users back to the top of the page. Making
navigation easier and more user friendly.

Change-Id: I80f415a60ea9c7df22276b464013e906e1efa511
Signed-off-by: Rodrigo Caballero <rodrigo.caballero.abraham@intel.com>

Former-commit-id: c4312b652c0d5feb3abfcc27e7e8c36ff7a63606