testcases: stub for x86 core on Quark SE
When running independent testcases that can run on any core, in
certain platforms using the Quark SE SOC, a stub is needed in the x86
to bring up the ARC core and pass along serial port messages. This
allows the indepent testcase to be ran in the ARC core without
modifications (the same way it runs on any other core).

There is a solution planned for the future that will allow this
workaround to be removed (without need to support backwards
compatibility).

Change-Id: I2a9b7e8797d03a5a2469d7b18a3025b79226f389
Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>

Former-commit-id: 8518d9cf81d2e8f76c1b174718f4796ccb1fed78